Updating test to only run on 1.6
